+++ R/forstringr-package.R
@@ -1,0 +1,11 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom dplyr %>%
+#' @importFrom dplyr across
+#' @importFrom dplyr mutate
+#' @importFrom stats na.omit
+#' @importFrom stringr str_to_title
+## usethis namespace: end
+NULL