+++ tests/testthat/test-str_title_case.R
@@ -1,0 +1,4 @@
+test_that("str_title_case() converts string to title case", {
+  expect_equal(str_title_case("the quick brown fox jumps over a lazy dog"), "The Quick Brown Fox Jumps over a Lazy Dog")
+})
+